Fixed the Clang build by fixing or supressing all the warnings.
Fixed types in console string manipulation to avoid errors/warnings with
signed to unsigned conversion.

Supressed a ton of warnings dealing with static/global destructors on
Mezzanine::String instances, these are either std::string or
Mezzanine::HashedString and will not cause issues regardless of destruction
ordering.

Reordered some padded items, surpressed weak vtables and padding warnings
around Mezzanine::testing::TestData because we simply don't care. Everything
will will be rebuilt on each target meaning the vtables will always match.
Trailing padding doesn't matter here (yet, because eventually this might be
profiled), because this is not performance sensitive code.

Move some variables in TestData implementation to an anonymous namespace to
futher reduce scope.

Supressed a ton of PugiXML warnings, This is a stopgap until we upgrade
versions.

Added some casts in situations where the type was known to be despite the
standard being a little stupid. (Why the hell does std::tolower return an int
and not a templated type?)

Fixed all the Include guards to surpress unused macro warnings, because every
use is flaw in our include chain.

Adjusted TestResultToString function to not have default case and rely on
compile-time type checking. The only time this function can be thrown is when
this function is fed garbage that has been cast or in the event of a hack. It
it the job of IO functions to filtering garbage when casting, and not this
functions job to stop hacks, so it was removed and appropriate GCC warning
supression added for the illusory bad return.

Fixed/supressed padding and weak-vtable warnings like I did on TestData for
the UnitTestGroup.