Reworked all the test macros to be thinner and call a function on the UnitTestGroup class.
I didn't redo TestTimed because that will require some re-thinking too. All of the redone macros are 3 or 4 lines long, much shorter than before.

I changed the TestResult enum to a class enum. This increasing verbosity which I suppose is bad, but it increases type safety and allowed me to
add an overload for operator << (ostream&, TestResult) that made sense.

I added a ton of tests to the Unit test suite for this.