Fixed tons of warnings with msvc.
Updated to new version of Jagati which suppress a ton of warnings globally
because ms throws them for the most inane things. There is a compiler
warning about ms optimizing code by inlining it. There is another for when
there are too many stack/automatically allocated floats (but not ints or
doubles).

At first the warnings for deleted operator= were nice, but then they were
thrown for test classes and could not be fixed with changes to the parent
class despite what the documentation says.