Many changes to help fix #4 and start on #7
I refactored include/ConsoleLogic.h and src/ConsoleLogic.cpp to make it more testable, simpler and include MEZZ_LIb. During this is when I discovered and attempted a fix for issue #4. I also removed sleep_for which was unimplemented.

I corrected some, but no all of the documentation in TestDataTools.h and added a new test macro TEST_STRING_CONTAINS which I back with a template in include/UnitTestGroup.h. I tested the macro in TestTests.

My temporary fix for #4 was adding a Name method to each of the tests classes in test/Boilerplate.h, test/Conversion.h, test/Test.h. I opened #8 to implement a longer term fix.