This is progress on the UnitTestGroup Refactor request by issue #8
Extracted OutputBufferGuard from UnitTestGroup.

Removed OverWriteResults enum and all uses of it. Replacing it with a single throw.

Fixed "is" -> "it" and missing MEZZ_LIB from previous PR.

This breaks subprocess output, but that is not yet tested or used. I will fix it and have unit test before this is done.