Updated the epsilon test to be better about testing floats.
On some platforms (MinGW in particular) double promotion is a warning, which is forced because "PreciseReal", aka double, was being used in the arithmetic.  This change will instead use the deduced "ExpectedResultsType" in the arithmetic which will prevent a promotion when floats are used.  As an additional layer of safety I've also added a static_assert to verify it is a float type (float, double, long double).